Fix overloading for nn and Functional modules
