+++ syft/frameworks/torch/nn/rnn.py
@@ -3,9 +3,8 @@ import torch
 from torch import nn
 from torch.nn import init
 
-import syft
-from syft.frameworks.torch.tensors.interpreters.precision import FixedPrecisionTensor
 from syft.frameworks.torch.tensors.interpreters.additive_shared import AdditiveSharingTensor
+from syft.frameworks.torch.tensors.interpreters.precision import FixedPrecisionTensor
 from syft.generic.pointers.pointer_tensor import PointerTensor
